move optimizer tests to core
